Update README with output variable types (#33)
The Maestro GH action returns variables that contain information about
the results of the run which can be very useful in CI. These variables
are noted in the README, but little information is given as to what
those variables will actually contain. This means devs need to run fake
CI runs, which take time and money, to see the type of output they're
going to get.

This PR simply clarifies what type the output variables will be, for
those types which are not simply strings, hopefully saving devs time in
debugging GH Actions workflows.