feat: add support for explicit branch specification (#57)
## The Problem

We have been running into an issue with the action where the `branch`
that is detected and used in the `UploadRequest` is not the branch we
are actually testing. This is because the branch is determined via
`getBranchName` which relies on the
`github.context.payload.pull_request` being present (or it falls back on
`github.context.ref`), neither of which are accurate for workflows
kicked off via `issue_comment`, `workflow_dispatch`, etc. if those
workflows intend to test code from a branch other than `main`.

## The Solution

This PR adds support for optionally specifying a `branch` input which
will be used in the downstream `UploadRequest`. By doing this, we
support branch-based cancelation which helps alleviate queue pressure.
Without this, if `Stop previous flows` is enabled and any of the
aforementioned workflow triggers are used, two unrelated flows may
cancel each other just by virtue of them both being kicked off/triggered
on `main`.

## Testing

There were no existing tests, so I added tests specifically for
`getBranchName` but nothing else. I also added a few other things that
we tend to use on our team for custom actions which may prove useful for
future improvements.

---------

Co-authored-by: Brandon Trautmann <8343465+btrautmann@users.noreply.github.com>
Co-authored-by: Dan Caseley <dan@caseley.me.uk>